# [Kernel] [FP8] Improve FP8 linear layer performance