--- a/vllm/_custom_ops.py
+++ b/vllm/_custom_ops.py
@@ -212,7 +212,9 @@ def scaled_fp8_quant(
     """
     if batch_dim_padding:
         shape = (max(batch_dim_padding, input.shape[0]), *input.shape[1:])
-        output = torch.empty(shape, device=input.device, dtype=torch.float8_e4m3fn)
+        output = torch.empty(shape,
+                             device=input.device,
+                             dtype=torch.float8_e4m3fn)
     else:
         output = torch.empty_like(input, dtype=torch.float8_e4m3fn)
     if scale is None: