--- a/vllm/model_executor/layers/quantization/fp8.py
+++ b/vllm/model_executor/layers/quantization/fp8.py
@@ -235,7 +235,8 @@ class Fp8LinearMethod(LinearMethodBase):
                                                layer.act_scale,
                                                batch_dim_padding=32)
 
-        # Fused GEMM_DQ
+        # Fused GEMM_DQ -- note we padded the input above because torch._scaled_mm
+        # is more performant for matrices with batch dimension at least 32.
         output, _ = torch._scaled_mm(
             qinput,
             layer.weight,