@robertgshaw2-neuralmagic 2024-05-08
kinda wild - I suspect we will be able to improve performance significantly with our kernels